docs: constant.js 변경 2023.08.16 1. constant 폴더 생성 2. Rename : constant.js -> home.js 3. profile.html 객체 상수화 예정
